Identifies biomolecules to be filtered in preparation for IMD-ANOVA.

The method identifies biomolecules to be filtered specifically according data requirements for running an ANOVA.

anova_filter(nonmiss_per_group, min_nonmiss_anova = 2, cname_id)

Arguments

nonmiss_per_group
a list of length two. The first element giving the total number of possible samples for each group. The second element giving a data.frame with the first column giving the biomolecule identifier and the second through kth columns giving the number of non-missing observations for each of the k groups. Usually the result of nonmissing_per_group

min_nonmiss_anova
the minimum number of nonmissing biomolecule values required, in each group, in order for the biomolecule to not be filtered. Must be greater than or equal to 2; default value is 2.

cname_id
character string specifying the name of the column containing the biomolecule identifiers in e_data and e_meta (if applicable).

Value

filter.peps a character vector of the biomolecules to be filtered out prior to ANOVA or IMD-ANOVA

Details

This function filters biomolecules that do not have at least min.nonmiss.allowed values per group, where groups are from group_designation.

See also

Examples

dontrun{
library(pmartRdata)
data("pep_object")
pep_object <- group_designation(pep_object, main_effects = "Condition")
filter_results <- anova_filter(nonmiss_per_group = nonmissing_per_group(omicsData = pep_object), cname_id = "Mass_Tag_ID")
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^