Convert Data to Appropriate pmartR Class

Converts a list object or several data.frames of isobaric peptide-level data to an object of the class 'isobaricpepData'. Objects of the class 'isobaricpepData' are lists with two obligatory components e_data and f_data. An optional list component e_meta is used if analysis or visualization at other levels (e.g. protein) is also desired.

as.isobaricpepData(e_data, f_data, e_meta = NULL, edata_cname, fdata_cname,
  emeta_cname = NULL, exp_cname, channel_cname = NULL,
  refpool_channel = NULL, refpool_cname = NULL, refpool_notation = NULL,
  ...)

Arguments

e_data
a \(p \times n + 1\) data.frame of expression data, where \(p\) is the number of peptides observed and \(n\) is the number of samples (an additional peptide identifier/name column should also be present anywhere in the data.frame). Each row corresponds to data for each peptide. One column specifying a unique identifier for each peptide (row) must be present.

f_data
a data.frame with \(n\) rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.

e_meta
an optional data.frame with at least \(p\) rows. Each row corresponds to a peptide with one column giving peptide names (must be named the same as the column in e_data) and other columns giving meta information (e.g. mappings of peptides to proteins).

edata_cname
character string specifying the name of the column containing the peptide identifiers in e_data and e_meta (if applicable).

fdata_cname
character string specifying the name of the column containing the sample identifiers in f_data.

emeta_cname
character string specifying the name of the column containing the protein identifiers (or other mapping variable) in e_meta (if applicable). Defaults to NULL. If e_meta is NULL, then either do not specify emeta_cname or specify it as NULL. If e_meta is NULL, then specify emeta_cname as NULL.

exp_cname
character string specifying the name of the column containing the experiment/plate information in f_data.

channel_cname
optional character string specifying the name of the column containing the instrument channel a sample was run on in f_data. This argument is optional, see Details for how to specify information regarding reference pool samples.

refpool_channel
optional character string specifying which channel contained the reference pool sample, only used when this remains the same from experiment to experiment. This argument is optional, see Details for how to specify information regarding reference pool samples.

refpool_cname
optional character string specifying the name of the column containing information about which samples are reference samples in f_data. This argument is optional, see Details for how to specify information regarding reference pool samples.

refpool_notation
optional character string specifying the value in the refpool_channel column which denotes that a sample is a reference sample. This argument is optional, see Details for how to specify information regarding reference pool samples.

...
further arguments

Details

The class 'isobaricpepData' is meant to deal with peptide data generated on instruments where a reference pool for normalization is available (e.g. TMT, iTRAQ). In all cases you must specify exp_cname which gives the column name for the column in f_data containing information about which experiment/plate a sample was run on. There are two ways to specify the information needed for identifying reference samples which should be used for normalization:

specify channel_cname and refpool_channel. This should be used when the reference sample for each experiment/plate was always located in the same channel. Here channel_cname gives the column name for the column in f_data which gives information about which channel each sample was run on, and refpool_channel is a character string specifying the value in channel_colname that corresponds to the reference sample channel.

specify refpool_cname and refpool_notation. This should be used when the reference sample is not in a consistent channel across experiments/plates. Here, refpool_cname gives the name of the column in f_data which indicates whether a sample is a reference or not, and refpool_notation is a character string giving the value used to denote a reference sample in that column.

See examples below. If your data has already undergone normalization to the reference pool, you should speficy isobaric_norm = T.

Objects of class 'isobaricpepData' contain some attributes that are referenced by downstream functions. These attributes can be changed from their default value by manual specification. A list of these attributes as well as their default values are as follows:

data_scale

Scale of the data provided in	e_data

. Acceptable values are 'log2', 'log10', 'log', and 'abundance', which indicate data is log base 2, base 10, natural log transformed, and raw abundance, respectively. Default is 'abundance'.

data_norm	A logical argument, specifying whether the data has been normalized or not. Default value is FALSE.

isobaric_norm

A logical argument, specifying whether the data has been normalized to the approporiate reference pool sample or not. Default value is FALSE

norm_info	Default value is NULL. When a normalization is applied to the data, this becomes populated with a list containing the normalization function, normalization subset and subset parameters, the location and scale parameters used to normalize the data, and the location and scale parameters used to backtransform the data (if applicable).

data_types

Character string describing the type of data (e.g.'Positive ion'). Default value is NULL.

check.names	Logical defaults to TRUE. Indicates whether 'check.names' attribute of returned omicsData object is TRUE or FALSE.

Computed values included in the data_info attribute are as follows:

num_edata

The number of unique	edata_cname

entries.

num_miss_obs	The number of missing observations.

num_emeta

The number of unique	emeta_cname

entries.

prop_missing	The proportion of

e_data	values that are NA.

num_samps

The number of samples that make up the columns of	e_data

.

meta_info	A logical argument, specifying whether

e_meta	is provided.

See also

Examples

dontrun{
library(pmartRdata)
data("isobaric_edata")
data("isobaric_fdata")
data("isobaric_emeta")
mypepData <- as.isobaricpepData(e_data = isobaric_edata, e_meta = isobaric_emeta, f_data = isobaric_fdata, edata_cname = "Peptide", fdata_cname = "Sample", emeta_cname = "Mass_Tag_ID")

# Alternate specification #

}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^