Convert Data to Appropriate pmartR Class

Converts a list object or several data.frames of lipid-level data to an object of the class 'lipidData'. Objects of the class 'lipidData' are lists with two obligatory components e_data and f_data. An optional list component e_meta is used if analysis or visualization at other levels (e.g. lipid identification) is also desired.

as.lipidData(e_data, f_data, e_meta = NULL, edata_cname, fdata_cname,
  emeta_cname = NULL, ...)

Arguments

e_data
a \(p \times n + 1\) data.frame of expression data, where \(p\) is the number of lipids observed and \(n\) is the number of samples. Each row corresponds to data for each lipid. One column specifying a unique identifier for each lipid (row) must be present.

f_data
a data.frame with \(n\) rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.

e_meta
an optional data.frame with \(p\) rows. Each row corresponds to a lipid with one column giving lipid names (must be named the same as the column in e_data) and other columns giving meta information.

edata_cname
character string specifying the name of the column containing the lipid identifiers in e_data and e_meta (if applicable).

fdata_cname
character string specifying the name of the column containing the sample identifiers in f_data.

emeta_cname
character string specifying the name of the column containing the mapped identifiers in e_meta (if applicable). Defaults to NULL. If e_meta is NULL, then either do not specify emeta_cname or specify it as NULL. If e_meta is NULL, then specify emeta_cname as NULL.

...
further arguments

Details

Objects of class 'lipidData' contain some attributes that are referenced by downstream functions. These attributes can be changed from their default value by manual specification. A list of these attributes as well as their default values are as follows:

data_scale	Scale of the data provided in e_data. Acceptable values are 'log2', 'log10', 'log', and 'abundance', which indicate data is log base 2, base 10, natural log transformed, and raw abundance, respectively. Default values is 'abundance'.

data_norm	A logical argument, specifying whether the data has been normalized or not. Default value is FALSE.

norm_info	Default value is NULL. When a normalization is applied to the data, this becomes populated with a list containing the normalization function, normalization subset and subset parameters, the location and scale parameters used to normalize the data, and the location and scale parameters used to backtransform the data (if applicable).

data_types	Character string describing the type of data (e.g. 'Positive ion'). Default value is NULL.

check.names	Logical defaults to TRUE. Indicates whether 'check.names' attribute of returned omicsData object is TRUE or FALSE.

data_scale

Computed values included in the data_info attribute are as follows:

num_edata	The number of unique edata_cname entries.

num_miss_obs	The number of missing observations.

num_emeta	The number of unique emeta_cname entries.

prop_missing	The proportion of e_data values that are NA.

num_samps	The number of samples that make up the columns of e_data.

meta_info	A logical argument, specifying where the e_meta is provided.

num_edata

See also

Examples

dontrun{
library(pmartRdata)
data("lipid_edata")
data("lipid_fdata")
mylipidData <- as.lipidData(e_data = lipid_edata, f_data = lipid_fdata, edata_cname = "LipidCommonName", fdata_cname = "Sample_Name")
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^