Runs BP-Quant

Applies BP-Quant to a pepData object

bpquant(statRes, pepData, pi_not = 0.9, max_proteoforms = 5)

Arguments

statRes
an object of the class 'statRes'

pepData
is an omicsData object of the class 'pepData'

pi_not
is a numeric value between 0 and 1 indicating the background probability/frequency of a zero signature.

max_proteoforms
a numeric value, a maximum threshold for the number of possible proteoforms.

Value

a list of data frames, one for each unique protein. The data frames have three columns, "Protein", "Mass_Tag_ID" (which is molecule ID), and "ProteoformID". The class of this list is 'isoformRes'.

Details

The statRes object contains the signatures data frame, the pepData object is used for its e_meta data frame. Next the signatures data frame and e_meta are merged by their edata_cname (“Peptide”) columns, this new data frame called protein_sig_data will be input to bpquant_mod in a “foreach” statement. “Foreach” will subset protein_sig_data for each unique protein and apply bpquant_mod to each subset and store the results.

Examples

dontrun{
library(pmartR)
library(pmartRdata)

mypepData <- group_designation(omicsData = pep_object, main_effects = c("Condition"))
mypepData = edata_transform(mypepData, "log2")

imdanova_Filt <- imdanova_filter(omicsData = mypepData)
mypepData <- applyFilt(filter_object = imdanova_Filt, omicsData = mypepData, min_nonmiss_anova=2)

imd_anova_res <- imd_anova(omicsData = mypepData, test_method = 'comb', pval_adjust='bon')

result = bpquant(statRes = imd_anova_res, pepData = mypepData)

}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^