bpquant_mod function

The function is written to take input from one protein at a time and requires three inputs: protein_sig, pi_not and max_proteforms.

bpquant_mod(protein_sig, pi_not, max_proteoforms)

Arguments

protein_sig
is a matrix or data.frame with p rows and n columns, where p is the number of peptides mapped to the protein of interest and n is the number of tests conducted to generate signatures made up of values 0, 1, and -1.

pi_not
is a numeric value between 0 and 1 indicating the background probability/frequency of a zero signature.

max_proteoforms
a numeric value, a maximum threshold for the number of possible proteoforms.

Value

a list of five items: num_proteoforms, unique_sigs, proteoform_configs, post_prob and peptide_idx

Details

num_proteoforms	the number of proteoforms as identified by bpquant

unique_sigs	matrix of unique signatures observed

proteoform_configs	matrix of 0/1 values indicating scenarios of proteoform absence/presence scenarios

post_prob	vector of posterior probabilities corresponding to each proteoform configuration in "proteoform_configs"

peptide_idx	vector of 0, 1, 2, . . . values indicating which proteoform each peptide belongs to

num_proteoforms

Examples

dontrun{
cur_protein = subset(protein_sig_data, Protein == "A")
cur_protein_sigs = cur_protein[,-(1:2)]

results = bpquant_mod(protein_sig = cur_protein_sigs, pi_not = .9, max_proteoforms = 5)
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^