Returns data frame with comparisons to be made

The method creates a data frame containing the comparisons to be made when performing differential statistics.

create_comparisonDF(comp_type, omicsData, control_group = NULL)

Arguments

comp_type

string specifying either "control" or "pairwise". Specifying "control" indicates that all other groups are to be compared to a single control group. Specifying "pairwise" indicates that all pairwise comparisons are to be made.

omicsData

A pmartR data object of any class, which has a `group_df` attribute created by the `group_designation()` function

control_group

string indicating the group to use for the control group. Only required when comp_type="control".

Value

data frame with columns for Test and Control. Each row corresponds to a comparison of interest.

Details

This function takes in the omicsData and type of comparison, and returns a data frame where each row corresponds to a comparison of interest.

See also

Author

Kelly Stratton

Examples

library(pmartRdata)
mymetab <- group_designation(omicsData = metab_object, main_effects = "Phenotype")
create_comparisonDF(comp_type = "pairwise", omicsData = mymetab)
#>         Test    Control
#> 1 Phenotype3 Phenotype2
#> 2 Phenotype1 Phenotype2
#> 3 Phenotype1 Phenotype3

create_comparisonDF(comp_type = "control", omicsData = mymetab, control_group = "Phenotype1")
#>         Test    Control
#> 1 Phenotype2 Phenotype1
#> 2 Phenotype3 Phenotype1