Pooled Coefficient of Variation (CV) Filter Object

A pooled CV is calculated for each biomolecule.

cv_filter(omicsData, use_groups = TRUE)

Arguments

omicsData

an object of class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData' created by as.pepData, as.proData, as.metabData, as.lipidData, or as.nmrData, respectively. Note, if group_designation has not been run, the CV is calculated based on all samples for each biomolecule.

use_groups

logical indicator for whether to utilize group information from group_designation when calculating the CV. Defaults to TRUE. If use_groups is set to TRUE but group_designation has not been run on the omicsData object, use_groups will be treated as FALSE.

Value

An S3 object of class 'cvFilt' giving the pooled CV for each biomolecule and additional information used for plotting a data.frame with a column giving the biomolecule name and a column giving the pooled CV value.

Details

For each biomolecule, the CV of each group is calculated as the standard deviation divided by the mean, excluding missing values. A pooled CV estimate is then calculated based on the methods of Ahmed (1995). Any groups consisting of a single sample are excluded from the CV calculation, and thus, from the cv_filter result. If group_designation has not been run on the omicsData object, all samples are considered to belong to the same group.

References

Ahmed, S.E. (1995). A pooling methodology for coefficient of variation. The Indian Journal of Statistics. 57: 57-75.

Author

Lisa Bramer, Kelly Stratton

Examples

library(pmartRdata)
mypep <- group_designation(omicsData = pep_object, 
                           main_effects = "Phenotype")
to_filter <- cv_filter(omicsData = mypep, use_groups = TRUE)
summary(to_filter, cv_threshold = 30)