Reduce Dimension of Data for Exploratory Data Analysis

Calculate principal components using projection pursuit estimation, which implements an expectation-maximization (EM) estimation algorithm when data is missing.

dim_reduction(omicsData, k = 2)

Arguments

omicsData
an object of the class 'pepdata', 'prodata', 'metabData', or 'lipidData' usually created by as.pepData, as.proData, as.metabData, or as.lipidData, respectively.

k
integer number of principal components to return. Defaults to 2.

Value

a data.frame with first k principal component scores, sample identifiers, and group membership for each sample (if group designation was previously run on the data).

Details

Any biomoleculs seen in only one sample or with a variance less than 1E-6 across all samples are not included in the PCA calculations. This function leverages code from pca.

References

Redestig, H., Stacklies, W., Scholz, M., Selbig, J., & Walther, D. (2007). pcaMethods - a bioconductor package providing PCA methods for incomplete data. Bioinformatics. 23(9): 1164-7.

Examples

dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object <- edata_transform(omicsData = lipid_object, data_scale="log2")
lipid_object <- group_designation(omicsData = lipid_object, main_effects = "Condition")
pca_lipids <- dim_reduction(omicsData = lipid_object)
plot(pca_lipids)
summary(pca_lipids)
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^