Creates a list of six Data Frames, one for each summarizing metric

This function takes in an omicsData object and returns a summary of the e_data component. The six summarizing metrics include the mean, standard deviation, median, percent observed, minimum, and maximum.

edata_summary(omicsData, by, groupvar = NULL)

Arguments

omicsData
an object of the class 'lipidData', 'metabData', 'pepData', or 'proData', usually created by as.lipidData, as.metabData, as.pepData, or as.proData, respectively.

by
a character string indicating whether summarizing metrics will be applied by 'sample' or by 'molecule

groupvar
a character vector with no more than two variable names that should be used to determine group membership of samples. The variable name must match a column name from f_data. Defaults to NULL, in which case group_DF attribute will be used.

Value

A list of six data frames, which are the results of applying the metrics (mean, standard deviation, median, percent observed, minimum and maximum) to omicsData$e_data

Details

If groupvar is NULL and group_designation has not been applied to omicsData, then the metrics will be applied to each column of e_data (when by = 'sample) or to each row of e_data (when by = 'molecule'). When groupvar is provided, it must match a column name from f_data, this column of f_data is used to group e_data in order to apply the metrics.

Examples

dontrun{
library(pmartRdata)
data(lipid_object)
lipid_object2 <- edata_transform(omicsData = lipid_object, data_scale = "log2")
lipid_object2 <- group_designation(omicsData = lipid_object, main_effects = "Condition")
edata_summary(omicsData = lipid_object2, by = "sample", groupvar = NULL)
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^