Apply a Transformation to the Data

This function applies a transformation to the e_data element of omicsData

edata_transform(omicsData, data_scale)

# S3 method for pepData
edata_transform(omicsData, data_scale)

# S3 method for proData
edata_transform(omicsData, data_scale)

# S3 method for metabData
edata_transform(omicsData, data_scale)

# S3 method for lipidData
edata_transform(omicsData, data_scale)

Arguments

omicsData
an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by as.pepData, as.proData, as.metabData, or as.lipidData, respectively.

data_scale
a character string indicating the type of transformation to be applied to the data. Valid values are: 'log2', 'log', 'log10', or 'abundance'. A value of 'abundance' indicates the data has previously undergone one of the log transformations and should be transformed back to raw values with no transformation applied.

Value

data object of the same class as omicsData

Details

This function is intended to be used before analysis of the data begins. Data are typically analyzed on a log scale.

Examples

dontrun{
library(pmartRdata)
data(metab_object)
metab_object2 <- edata_transform(omicsData = metab_object, data_scale="log2")
attr(metab_object2, "data_info")$data_scale
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^