Group comparisons for the anova test

Takes the results of anova_test() and returns group comparison p-values

group_comparison_anova(
  data,
  groupData,
  comparisons,
  Xfull,
  Xred,
  anova_results_full,
  beta_to_mu_full,
  beta_to_mu_red
)

Arguments

data

The expression values without the id column

groupData

data frame that assigns sample names to groups

comparisons

dataframe that defines the comparsions of interest

Xfull

design matrix for the full model with interaction terms between the main effects

Xred

design matrix for the reduced model with no interaction terms between the main effects

anova_results_full

results of the pmartR::anova_test() function

beta_to_mu_full

matrix that maps the beta coefficients to the group means for the full model

beta_to_mu_red

matrix that maps the beta coefficients to the group means for the reduced model

Value

A data.frame containing the p-values from the group comparisons.

Author

Bryan Stanfill, Daniel Claborne