Tests for a qualitative and quantiative difference between groups using IMD and ANOVA, respectively

This is IMD-ANOVA test proposed in Webb-Robertson et al. (2010).

imd_anova(omicsData, comparisons = NULL, test_method, pval_adjust = "none",
  pval_thresh = 0.05, covariates = NULL, paired = FALSE,
  equal_var = TRUE)

Arguments

omicsData
A pmartR data object of any class, which has a `group_df` attribute that is usually created by the `group_designation()` function

comparisons
data.frame with columns for "Control" and "Test" containing the different comparisons of interest. Comparisons will be made between the Test and the corresponding Control If left NULL, then all pairwise comparisons are executed.

test_method
character string specifying the filter method to use: "combined", "gtest", or "anova". "combined" implements both the gtest and anova filters.

pval_adjust
character vector specifying the type of multiple comparisons adjustment to implement via p.adjust. A NULL value corresponds to no adjustment. Valid options for ANOVA include: holm, hochberg, hommel, bonferroni, BH, BY, fdr, none. Valid options for g-test include: holm, bonferonni and none. See p.adjust for some details.

pval_thresh
numeric p-value threshold, below or equal to which peptides are considered differentially expressed. Defaults to 0.05

covariates
data.frame similar to groupData consisting of two columns: the sample ID variable (with names matching column names in omicsData$e_data) and a column containing the numeric or group data for each sample

paired
logical - should the data be paired or not; if TRUE the attribute "pairing" cannot be `NULL`

Value

a list of data.frame objects

Full_Results

Columns: e_data cname, group counts, group means, ANOVA p-values, IMD p-values, fold change estimates, fold change significance flags

Flags (signatures)

Indicator of statistical significance for one (+/-1 for ANOVA, +/-2 for g-test) or neither (0) test

p-values

p-values for pairwise comparisons adjusted for within, e.g., peptide multiple comparisons only (not across peptides yet)

groupData	Mapping between sample ID and group

References

Webb-Robertson, Bobbie-Jo M., et al. "Combined statistical analyses of peptide intensities and peptide occurrences improves identification of significant peptides from MS-based proteomics data." Journal of proteome research 9.11 (2010): 5748-5756.

Examples

dontrun{
library(pmartR)
library(pmartRdata)
#Transform the data
mypepData <- edata_transform(omicsData = pep_object, data_scale = "log2")

#Group the data by condition
mypepData <- group_designation(omicsData = mypepData, main_effects = c("Condition"))

#Apply the IMD ANOVA filter
imdanova_Filt <- imdanova_filter(omicsData = mypepData)
mypepData <- applyFilt(filter_object = imdanova_Filt, omicsData = mypepData, min_nonmiss_anova=2)

#Implement the IMD ANOVA method and compute all pairwise comparisons (i.e. leave the `comparisons` argument NULL)
anova_res <- imd_anova(omicsData = mypepData, test_method = 'anova')
imd_res <- imd_anova(omicsData = mypepData, test_method = 'gtest')
imd_anova_res <- imd_anova(omicsData = mypepData, test_method = 'comb', pval_adjust='bon')

#Test with really big dataset
library(OvarianPepdataBP)
tcga_ovarian_pepdata_bp <- as.pepData(e_data = tcga_ovarian_pepdata_bp$e_data, f_data = tcga_ovarian_pepdata_bp$f_data, e_meta = tcga_ovarian_pepdata_bp$e_meta)
tcga_ovarian_pepdata_bp <- group_designation(omicsData = tcga_ovarian_pepdata_bp, main_effects = c("race"))
ovarian_res <- imd_anova(omicsData = tcga_ovarian_pepdata_bp, test_method = 'anova')
#Tukey adjustment is super slow right now because "ptukey" is super slow, not sure how to fix that
ovarian_res_holm <- imd_anova(omicsData = tcga_ovarian_pepdata_bp, pval_adjust = 'holm', test_method='gtest')
#Dunnett adjustment, should give an error because dunnett correction shouldn't be applied for all pairwise comparisons
ovarian_res_dunnett <- imd_anova(omicsData = tcga_ovarian_pepdata_bp, pval_adjust = 'dunnett', test_method='combined')

#Test really big dataset, two factors
tcga_ovarian_pepdata_bp <- group_designation(omicsData = tcga_ovarian_pepdata_bp, main_effects = c("vital_status","neoplasm_histologic_grade"))
ovarian_res_twofac <- imd_anova(omicsData = tcga_ovarian_pepdata_bp, test_method='comb')

#Same but only test main effects (Dead vs Alive, G2 vs G3)
comp_df <- data.frame(Control=c("Alive","G2"), Test=c("Dead","G3"))
ovarian_res_twofac_main_effects <- imd_anova(omicsData = tcga_ovarian_pepdata_bp, comparisons = comp_df, test_method='comb')
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^