Produces a volcano plot from statRes object

This function creates a fold change vs t-Test p-value volcano plot from a statRes object

missingval_volcanoplot(statRes, comparison, x_lab = NULL, ...)

Arguments

statRes
is an object of class 'statRes'

comparison
can either be a character string name of the comparison to plot, or an integer index refering to the comparisons attribute vector

x_lab
character string to be used for x-axis label. Defaults to NULL

...
further arguments

y_lab

character string to be used for y-axis label. Defaults to NULL

title_plot

character string to be used for the plot title. Defaults to NULL.

legend_title

character string to be used for legend_title label. Defaults to NULL

title_size

integer value specifying the font size for the plot title. Default is 14.

x_lab_size

integer value indicating the font size for the x-axis. Defaults to 11.

y_lab_size

integer value indicating the font size for the y-axis. Defaults to 11.

bw_theme

logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used.

vlines

The x coordinate (integer in absolute value) where to draw vertical lines, defaults to NULL

pvalue_threshold

numeric value, draws horizontal line at value, defaults to NULL

palette

is a character string indicating the name of the RColorBrewer palette to use; "YlOrRd", "YlOrBr", "YlGnBu", "YlGn", "Reds","RdPu", "Purples", "PuRd", "PuBuGn", "PuBu", "OrRd","Oranges", "Greys", "Greens", "GnBu", "BuPu","BuGn","Blues", "Set3", "Set2", "Set1", "Pastel2", "Pastel1", "Paired", "Dark2", "Accent", "Spectral", "RdYlGn", "RdYlBu", "RdGy", "RdBu", "PuOr","PRGn", "PiYG", "BrBG"

x_lab_angle

integer value indicating the angle of x-axis labels

coordinate_flip

logical indicates whether to flip cartesian coordinates so that horizontal becomes vertical and vise versa, defaults to false

Value

plots ggplot2 object

Details

This function takes in a statRes object and plots log2 fold change vs -log10 t-Test P-values for a comparison from attr(statRes, “comparison”), ‘comparison’ parameter could be either a string or an integer refereing to an index of statRes comparisons

Examples

dontrun{
library(pmartR)
library(pmartRdata)

myproData <- group_designation(omicsData = pro_object, main_effects = c("Condition"))

imdanova_Filt <- imdanova_filter(omicsData = myproData)
myproData <- applyFilt(filter_object = imdanova_Filt, omicsData = myproData, min_nonmiss_anova=2)

statRes_object <- imd_anova(omicsData = myproData, test_method = 'comb', pval_adjust='bon')

missingval_volcanoplot(statRes_object, comparison = 2, palette = "Set1")
missingval_volcanoplot(statRes_object, comparison = 3, pvalue_threshold = .005, palette = "Dark2", x_lab_angle = 45)

}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^