Examine and Apply Isobaric Normalization

Examine reference pool samples and apply normalization of study samples to their corresponding reference pool sample

normalize_isobaric(
  omicsData,
  exp_cname = NULL,
  apply_norm = FALSE,
  channel_cname = NULL,
  refpool_channel = NULL,
  refpool_cname = NULL,
  refpool_notation = NULL
)

Arguments

omicsData

an object of the class 'isobaricpepData'

exp_cname

character string specifying the name of the column containing the experiment/plate information in f_data

apply_norm

logical, indicates whether normalization should be applied to omicsData$e_data

channel_cname

optional character string specifying the name of the column containing the instrument channel a sample was run on in f_data. This argument is optional. See Details for how to specify information regarding reference pool samples. If using this argument, the 'refpool_channel' argument must also be specified; in this case, 'refpool_cname' and 'refpool_notation' should not be specified.

refpool_channel

optional character string specifying which channel contains the reference pool sample. Only used when this is the same from experiment to experiment. This argument is optional. See Details for how to specify information regarding reference pool samples. If using this argument, the 'channel_cname' argument must also be specified; in this case, 'refpool_cname' and 'refpool_notation' should not be specified.

refpool_cname

optional character string specifying the name of the column containing information about which samples are reference samples in f_data. This argument is optional. see Details for how to specify information regarding reference pool samples. If using this argument, the 'refpool_notation' argument must also be specified; in this case, 'channel_cname' and 'refpool_channel' should not be specified.

refpool_notation

optional character string specifying the value in the refpool_channel column which denotes that a sample is a reference sample. This argument is optional. See Details for how to specify information regarding reference pool samples. If using this argument, the 'refpool_cname' argument must also be specified; in this case, 'channel_cname' and 'refpool_channel' should not be specified.

Value

If apply_norm = TRUE, an object of class 'isobaricpepData', normalized to reference pool, and with the attribute 'isobaric_info' updated to include information about the reference pool samples and the normalization procedure. Otherwise an object of class 'isobaricnormRes' containing similar information about the normalization process

Details

There are two ways to specify the information needed for identifying reference samples which should be used for normalization:

specify channel_cname and refpool_channel. This should be used when the reference sample for each experiment/plate was always located in the same channel. Here channel_cname gives the column name for the column in f_data which gives information about which channel each sample was run on, and refpool_channel is a character string specifying the value in channel_colname that corresponds to the reference sample channel.

specify refpool_cname and refpool_notation. This should be used when the reference sample is not in a consistent channel across experiments/plates. Here, refpool_cname gives the name of the column in f_data which indicates whether a sample is a reference or not, and refpool_notation is a character string giving the value used to denote a reference sample in that column.

In both cases you must specify exp_cname which gives the column name for the column in f_data containing information about which experiment/plate a sample was run on.

Examples

library(pmartRdata)

myiso <- edata_transform(isobaric_object, "log2")

# Don't apply the normalization quite yet;
# can use summary() and plot() to view reference pool samples
myiso_refpools <- normalize_isobaric(
  omicsData = myiso, exp_cname = "Plex",
  apply_norm = FALSE,
  refpool_cname = "Virus",
  refpool_notation = "Pool"
)
summary(myiso_refpools)

# Now apply the normalization;
# can use plot() to view the study samples after reference pool normalization
myiso_norm <- normalize_isobaric(
  omicsData = myiso, exp_cname = "Plex",
  apply_norm = TRUE,
  refpool_cname = "Virus",
  refpool_notation = "Pool"
)