Plots an object of class naRes

For plotting an S3 object of type 'naRes':

# S3 method for naRes
plot(naRes_object, type, x_lab = NULL, ...)

Arguments

naRes_object
a list of two data frames, one contains NA values by sample, the second contains NA values by molecule

type
is for specifying plot type, there are three options, 'bySample': plots missing values per sample, 'byMolecule': plots histogram for the number of missing values per molecule and 'Both': displays both 'bySample' and 'byMolecule' plots.

x_lab
character string to be used for x-axis label. Defaults to NULL

...
further arguments

y_lab

character string to be used for y-axis label. Defaults to NULL

title_plot

character string to be used for the plot title. Defaults to NULL.

legend_title

character string to be used for legend_title label. Defaults to NULL

title_size

integer value specifying the font size for the plot title. Default is 14.

x_lab_size

integer value indicating the font size for the x-axis. Defaults to 11.

y_lab_size

integer value indicating the font size for the y-axis. Defaults to 11.

bar_width

integer value indicating the bar width in a barplot (when type = "bySample"). Defaults to .8.

binwidth

integer value indicating the bin width in a histogram (when type = "byMolecule"). Defaults to 1.

bw_theme

logical indicator of whether to use the "theme_bw". Defaults to FALSE, in which case the ggplot2 default theme is used.

palette

palette is a character string indicating the name of the RColorBrewer palette to use; "YlOrRd", "YlOrBr", "YlGnBu", "YlGn", "Reds","RdPu", "Purples", "PuRd", "PuBuGn", "PuBu", "OrRd","Oranges", "Greys", "Greens", "GnBu", "BuPu","BuGn","Blues", "Set3", "Set2", "Set1", "Pastel2", "Pastel1", "Paired", "Dark2", "Accent", "Spectral", "RdYlGn", "RdYlBu", "RdGy", "RdBu", "PuOr","PRGn", "PiYG", "BrBG"

x_lab_angle

integer value indicating the angle of x-axis labels

coordinate_flip

logical indicates whether to flip cartesian coordinates so that horizontal becomes vertical and vise versa, defaults to false

use_VizSampNames

logical indicates whether to use custom sample names

Value

plots ggplot2 object

Details

This function takes in an object of class naRes and creates either a bar chart or histogram of missing values. When parameter type = "bySample", a "sample name" by "missing values count" bar chart is returned. When parameter type = "byMolecule" a "number of missing values per molecule" by "Molecules" histogram is returned. If type is set to "Both", a bar chart and histogram will be returned.

Examples

dontrun{
library(pmartRdata)
data("lipid_object")
result<- missingval_result(lipid_object)
plot(result, type = "bySample", x_lab_angle = 50, palette = "Set1")

plot(result, type = "Both", palette = "Set2")
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^