Protein Quantification

This function takes in a pepData object, method (quantification method, mean, median or rrollup), and the optional argument isoformRes (defaults to NULL). An object of the class 'proData' is returned.

protein_quant(
  pepData,
  method,
  isoformRes = NULL,
  qrollup_thresh = NULL,
  single_pep = FALSE,
  single_observation = FALSE,
  combine_fn = "median",
  parallel = TRUE,
  emeta_cols = NULL
)

Arguments

pepData

an omicsData object of the class 'pepData'

method

character string specifying one of four protein quantification methods, 'rollup', 'rrollup', 'qrollup' and 'zrollup'

isoformRes

list of data frames, the result of applying the 'bpquant' function to original pepData object. Defaults to NULL.

qrollup_thresh

numeric value; is the peptide abundance cutoff value. Is an argument to qrollup function.

single_pep

logical indicating whether or not to remove proteins that have just a single peptide mapping to them, defaults to FALSE.

single_observation

logical indicating whether or not to remove peptides that have just a single observation, defaults to FALSE.

combine_fn

character string specifying either be 'mean' or 'median'

parallel

logical indicating whether or not to use "doParallel" loop in applying rollup functions. Defaults to TRUE. Is an argument of rrollup, qrollup and zrollup functions.

emeta_cols

character vector indicating additional columns of e_meta that should be kept after rolling up to the protein level. The default, NULL, only keeps the column containing the mapping variable along with the new columns created (peps_per_pro and n_peps_used).

Value

omicsData object of the class 'proData'

Details

If isoformRes is provided then, a temporary pepData object is formed using the isoformRes information as the e_meta component and the original pepData object will be used for e_data and f_data components. The emeta_cname for the temporary pepData object will be the 'protein_isoform' column of isoformRes. Then one of the three 'method' functions can be applied to the temporary pepData object to return a proData object. If isofromRes is left NULL, then depending on the input for 'method', the correct 'method' function is applied directly to the input pepData object and a proData object is returned.

References

Webb-Robertson, B.-J. M., Matzke, M. M., Datta, S., Payne, S. H., Kang, J., Bramer, L. M., ... Waters, K. M. (2014). Bayesian Proteoform Modeling Improves Protein Quantification of Global Proteomic Measurements. Molecular & Cellular Proteomics.: MCP, 13(12), 3639-3646.

Examples

if (FALSE) {
library(pmartRdata)

mypepData <- group_designation(omicsData = pep_object, main_effects = c("Phenotype"))
mypepData = edata_transform(omicsData = mypepData, "log2")

imdanova_Filt <- imdanova_filter(omicsData = mypepData)
mypepData <- applyFilt(filter_object = imdanova_Filt, omicsData = mypepData, min_nonmiss_anova=2)

imd_anova_res <- imd_anova(omicsData = mypepData, test_method = 'comb', pval_adjust_a_multcomp ='bon', pval_adjust_g_multcomp ='bon')

isoformRes = bpquant(statRes = imd_anova_res, pepData = mypepData)

#case where isoformRes is NULL:
results<- protein_quant(pepData = mypepData, method = 'rollup', combine_fn = 'median', isoformRes = NULL)

#case where isoformRes is provided:
# results2 = protein_quant(pepData = mypepData, method = 'rollup', combine_fn = 'mean', isoformRes = isoformRes)
}