Creates a data frame displaying multiple metrics

This function takes in an object of class 'dataRes' and returns a data frame displaying a combination of metrics. The six summarizing metrics include, mean, standard deviation, median, percent observed, minimum, and maximum.

report_dataRes(dataRes, minmax = FALSE, digits = 2)

Arguments

minmax

logical specifying whether or not to include minimum and maximum data in the returned data frame. Defaults to FALSE.

digits

integer indicating the number of decimal places to round

statRes

an object of the class 'dataRes', created by edata_summary.

Value

prints a data frame

Details

When creating the 'dataRes' object via edata_summary, if the 'by' argument is set to 'sample', then the 'groupvar' argument must be NULL

Examples

library(pmartRdata)
mylipid <- edata_transform(omicsData = lipid_neg_object, data_scale = "log2")

dataRes_sample <- edata_summary(omicsData = mylipid, groupvar = NULL, by = "sample")
my_output <- report_dataRes(dataRes_sample)