Conversion between log2(RMD) and p-value

This function provides a conversion between the log base 2 robust Mahalanobis distance value and p-value for output from the rmd_runs function

rmd_conversion(log2rmd = NULL, pval = NULL, df)

Arguments

log2rmd

numeric log base 2 transformed robust Mahalanobis distance value

pval

numeric p-value associated with rmd_runs algorithm

df

integer value specifying the degrees of freedom associated with the test, which should be equal to the number of metrics used in rmd_runs

Value

The function returns the corresponding p-value or log base 2 robust Mahalanobis when the other parameter is specified.

Details

Only one of log2rmd and pval should be provided. The input not provided will be solved for based on the provided input.

Author

Lisa Bramer

Examples

library(pmartRdata)
mymetab <- edata_transform(
  omicsData = metab_object,
  data_scale = "log2"
)
mymetab <- group_designation(
  omicsData = mymetab,
  main_effects = "Phenotype"
)
rmd_results <- rmd_filter(
  omicsData = mymetab,
  metrics = c("MAD", "Skewness", "Correlation")
)
rmd_conversion(log2rmd = rmd_results$Log2.md, df = 3)

rmd_conversion(pval = .0001, df = 3)
rmd_conversion(log2rmd = 4.5, df = 3)