RMD Runs

The method computes a robust Mahalanobis distance that can be mapped to a p-value and used to identify outlying samples

rmd_filter(omicsData, ...)

# S3 method for pepData
rmd_filter(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation", "Proportion_Missing"))

# S3 method for proData
rmd_filter(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation", "Proportion_Missing"))

# S3 method for lipidData
rmd_filter(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation"))

# S3 method for metabData
rmd_filter(omicsData, metrics = c("MAD", "Kurtosis",
  "Skewness", "Correlation"))

Arguments

omicsData
an object of the class 'pepData', 'proData', 'metabData', or 'lipidData', usually created by as.pepData, as.proData, as.metabData, or as.lipidData, respectively.

...
further arguments

Value

a data.frame containing columns for the Sample ID, log2 robust Mahalanobis distance, p-values, and robust Mahalanobis distance

Details

The metrics on which the log2 robust Mahalanobis distance is based can be specified using the metrics argument.

pepData

For pepData objects, all five of the metrics "MAD", "Kurtosis", "Skewness", "Correlation", "Proportion_Missing" may be used (this is the default).

proData	For proData objects, all five of the metrics "MAD", "Kurtosis", "Skewness", "Correlation", "Proportion_Missing" may be used (this is the default).

metabData

For metabData objects, the use of "Proportion_Missing" is discouraged due to the general lack of missing data in metabolomics datasets (the default behavior omits "Proportion_Missing" from the metrics).

lipidData	For lipidData objects, , the use of "Proportion_Missing" is discouraged due to the general lack of missing data in metabolomics datasets (the default behavior omits "Proportion_Missing" from the metrics).

References

Matzke, M., Waters, K., Metz, T., Jacobs, J., Sims, A., Baric, R., Pounds, J., and Webb-Robertson, B.J. (2011), Improved quality control processing of peptide-centric LC-MS proteomics data. Bioinformatics. 27(20): 2866-2872.

Examples

dontrun{
library(pmartRdata)
data(metab_object)
metab_object2 <- edata_transform(omicsData = metab_object, data_scale = "log2")
metab_object3 <- group_designation(omicsData = metab_object2, main_effects = "Condition")
rmd_results <- rmd_filter(omicsData = metab_object3, metrics=c("MAD", "Skewness", "Correlation"))
rmd_results <- rmd_filter(omicsData = metab_object2)

data(pep_pepData)
pep_pepData2 <- edata_transform(omicsData = pep_object, data_scale = "log2")
pep_pepData3 <- group_designation(omicsData = pep_pepData2, main_effects = "Condition")
rmd_results <- rmd_filter(omicsData = pep_pepData3)
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^