Fold change barplots for statres objects

Plots a bar-chart with bar heights indicating the number of significant biomolecules, grouped by test type and fold change direction.

statres_barplot(
  x,
  stacked,
  fc_colors,
  text_size,
  display_count,
  x_lab,
  y_lab,
  title_lab,
  legend_lab,
  free_y_axis
)

Arguments

x, stacked, fc_colors

passed from pmartR::plot.statRes()

text_size

An integer specifying the size of the text (number of non-missing values) within the plot. The default is 3.

display_count

logical value. Indicates whether the non-missing counts will be displayed on the bar plot. The default is TRUE.

x_lab

character string specifying the x-axis label.

y_lab

character string specifying the y-axis label.

title_lab

character string specifying the plot title.

legend_lab

character string specifying the legend title.

Value

`ggplot` object - barplot.