Produce a basic summary of an pmartR S3 Object

This function will provide basic summary statistics for objects from the MSomicsR package.

# S3 method for pepData
summary(omicsData)

# S3 method for proData
summary(omicsData)

# S3 method for lipidData
summary(omicsData)

# S3 method for metabData
summary(omicsData)

# S3 method for corRes
summary(corRes_object)

# S3 method for dimRes
summary(dimRes_object)

# S3 method for moleculeFilt
summary(filter_object, min_num = NULL)

# S3 method for proteomicsFilt
summary(filter_object, min_num_peps = NULL,
  degen_peps = FALSE)

# S3 method for imdanovaFilt
summary(filter_object, min_nonmiss_anova = NULL,
  min_nonmiss_gtest = NULL)

# S3 method for rmdFilt
summary(filter_object, pvalue_threshold = NULL)

# S3 method for cvFilt
summary(filter_object, cv_threshold = NULL)

# S3 method for customFilt
summary(filter_object)

# S3 method for normRes
summary(omicsNorm)

Arguments

omicsData
an object of the class 'lipidData', 'metabData', 'pepData', or 'proData' usually created by as.lipidData, as.metabData, as.pepData, or as.proData, respectively.

corRes_object
an object of class corRes. A correlation matrix of all samples.

dimRes_object
an object of class dimRes. A list containing sample identifiers and the principle components scores.

filter_object
S3 object of class 'moleculeFilt', 'proteomicsFilt', or 'imdanovaFilt' created by molecule_filter, proteomics_filter, or imdanova_filter, respectively.

min_num
an integer value specifying the minimum number of times each feature must be observed across all samples. Default value is 2.

min_num_peps
an optional integer value between 1 and the maximum number of peptides that map to a protein in the data. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than min_num_peps mapping to it will be returned as a protein that should be filtered. Default value is NULL.

degen_peps
logical indicator of whether to filter out degenerate peptides (TRUE) or not (FALSE). Default value is FALSE.

min_nonmiss_anova
the minimum number of non-missing feature values allowed per group for anova_filter. Suggested value is 2.

min_nonmiss_gtest
the minimum number of non-missing feature values allowed per group for gtest_filter. Suggested value is 3.

pvalue_threshold
A threshold for the Robust Mahalanobis Distance (RMD) p-value. All samples below the threshold will be filtered out. Default value is NULL.

cv_threshold
CV values above cv_threshold are filtered out. Default value is NULL.

Value

a summary table for the MSomicsR object. If assigned to a variable, the elements of the summary table are saved in a list format.

Examples

dontrun{
library(pmartRdata)
data("pep_object")
summary(pep_object)
to_filter <- molecule_filter(pep_object)
summary(to_filter)
summary(to_filter, min_num = 4)
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^