Proteomics Filter Summary

Provide summary of a proteomicsFilt S3 object

# S3 method for proteomicsFilt
summary(object, min_num_peps = NULL, degen_peps = FALSE, ...)

Arguments

object

S3 object of class 'proteomicsFilt' created by proteomics_filter.

min_num_peps

optional integer value between 1 and the maximum number of peptides that map to a protein in the data. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than min_num_peps mapping to it will be returned as a protein that should be filtered. Default value is NULL.

degen_peps

logical indicator of whether to filter out 'degenerate' or 'redundant' peptides (i.e. peptides mapping to multiple proteins) (TRUE) or not (FALSE). Default value is FALSE.

...

further arguments passed to or from other methods

Value

a summary table giving the number of Observed Proteins per Peptide and number of Observed Peptides per Protein. If min_num_peps is specified and/or degen_peps is TRUE, the number of biomolecules to be filtered with the specified threshold(s) are reported.

See also

Author

Lisa Bramer

Examples

library(pmartRdata)
myfilt <- proteomics_filter(omicsData = pep_object)
summary(myfilt, degen_peps = TRUE) # there are no degenerate peptides to filter out
summary(myfilt, min_num_peps = 2)