Applies zrollup function

This function applies the zrollup method to a pepData object for each unique protein and returns a proData object.

zrollup(pepData, combine_fn, parallel = TRUE)

Arguments

pepData

an omicsData object of class 'pepData'

combine_fn

logical indicating what combine_fn to use, defaults to median, other option is mean

parallel

logical indicating whether or not to use "doParallel" loop in applying zrollup function. Defaults to TRUE.

Value

an omicsData object of class 'proData'

Details

In the zrollup method, peptides are scaled as, pep_scaled = (pep - median)/sd, and protein abundance is set as the mean of these scaled peptides.

References

Polpitiya, A. D., Qian, W.-J., Jaitly, N., Petyuk, V. A., Adkins, J. N., Camp, D. G., ... Smith, R. D. (2008). DAnTE: a statistical tool for quantitative analysis of -omics data. Bioinformatics (Oxford, England), 24(13), 1556-1558.