Identifies biomolecules to be filtered in preparation for IMD-ANOVA.

The method identifies biomolecules to be filtered specifically according data requirements for running an ANOVA.

anova_filter(nonmiss_per_group, min_nonmiss_anova, comparisons = NULL)

Arguments

nonmiss_per_group

a list of length two. The first element giving the total number of possible samples for each group. The second element giving a data.frame with the first column giving the biomolecule identifier and the second through kth columns giving the number of non-missing observations for each of the k groups. Usually the result of nonmissing_per_group

min_nonmiss_anova

the minimum number of nonmissing biomolecule values required, in each group, in order for the biomolecule to not be filtered. Must be greater than or equal to 2; default value is 2.

comparisons

data.frame with columns for "Control" and "Test" containing the different comparisons of interest. Comparisons will be made between the Test and the corresponding Control. If left NULL, then all pairwise comparisons are executed.

Value

filter.peps a character vector of the biomolecules to be filtered out prior to ANOVA or IMD-ANOVA

Details

This function filters biomolecules that do not have at least min.nonmiss.allowed values per group, where groups are from group_designation.

See also

Author

Kelly Stratton