Apply a S3 filter object to a pmartR S3 object

This function takes a filter object of class 'cvFilt', 'rmdFilt', 'moleculeFilt', 'proteomicsFilt', 'imdanovaFilt', 'RNAFilt', 'totalCountFilt', or 'customFilt' and applies the filter to a dataset of pepData, proData, lipidData, metabData, nmrData or seqData.

applyFilt(filter_object, omicsData, ...)

# S3 method for moleculeFilt
applyFilt(filter_object, omicsData, min_num = 2)

# S3 method for totalCountFilt
applyFilt(filter_object, omicsData, min_count)

# S3 method for RNAFilt
applyFilt(filter_object, omicsData, min_nonzero = NULL, size_library = NULL)

# S3 method for cvFilt
applyFilt(filter_object, omicsData, cv_threshold = 150)

# S3 method for rmdFilt
applyFilt(
  filter_object,
  omicsData,
  pvalue_threshold = 1e-04,
  min_num_biomolecules = 50
)

# S3 method for proteomicsFilt
applyFilt(filter_object, omicsData, min_num_peps = NULL, redundancy = FALSE)

# S3 method for imdanovaFilt
applyFilt(
  filter_object,
  omicsData,
  comparisons = NULL,
  min_nonmiss_anova = NULL,
  min_nonmiss_gtest = NULL,
  remove_singleton_groups = TRUE
)

# S3 method for customFilt
applyFilt(filter_object, omicsData)

Arguments

filter_object

an object of the class 'cvFilt', 'proteomicsFilt', 'rmdFilt', 'moleculeFilt', 'imdanovaFilt', 'customFilt', 'RNAFilt', or 'totalCountFilt' created by cv_filter, proteomics_filter, rmd_filter, molecule_filter, imdanova_filter, custom_filter, RNA_filter, or total_count_filter, respectively.

omicsData

an object of the class pepData, proData, lipidData, metabData, nmrData, or seqData usually created by as.pepData, as.proData, as.lipidData, as.metabData, as.nmrData, or as.seqData, respectively.

...

further arguments

Value

An object of the class pepData, proData,

lipidData, metabData, nmrData, or seqData with specified cname_ids, edata_cnames, and emeta_cnames filtered out of the appropriate datasets.

Details

Further arguments can be specified depending on the class of the filter_object being applied.

For a filter_object of type 'moleculeFilt':

min_num	an integer value specifying the minimum number of times each biomolecule must be observed across all samples in order to retain the biomolecule. Default value is 2.

For a filter_object of type 'cvFilt':

cv_threshold	an integer value specifying the maximum coefficient of variation (CV) threshold for the biomolecules. Biomolecules with CV greater than this threshold will be filtered. Default threshold is 150.

For a filter_object of type 'rmdFilt':

pvalue_threshold	numeric value between 0 and 1, specifying the p-value below which samples will be removed from the dataset. Default is 0.001.

min_num_biomolecules	numeric value greater than 10 (preferably greater than 50) that specifies the minimum number of biomolecules that must be present in order to create an rmdFilt object. Using values less than 50 is not advised.

For a filter_object of type 'proteomicsFilt' either or both of the following can be applied:

min_num_peps	an optional integer value between 1 and the maximum number of peptides that map to a protein in omicsData. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than min_num_peps mapping to it will be removed from the dataset. Default value is NULL, meaning that this filter is not applied.

redundancy	logical indicator of whether to filter out degenerate/redundant peptides (peptides that map to more than one protein). Default value is FALSE.

For a filter_object of type 'imdanovaFilt':

min_nonmiss_anova	integer value specifying the minimum number of non-missing feature values allowed per group for anova_filter. Default value is 2.

min_nonmiss_gtest	integer value specifying the minimum number of non-missing feature values allowed per group for gtest_filter. Default value is 3.

For a filter_object of type 'totalCountFilt':

min_count	integer value specifying the minimum number of biomolecule counts observed across all samples in order for the biomolecule to be retained in the dataset. This filter is only applicable for seqData objects.

For a filter_object of type 'RNAFilt' either or both of the following can be applied:

min_nonzero	integer value specifying the minimum number of non-zero feature values per sample.

size_library	integer value or fraction between 0 and 1 specifying the minimum number of total reads per sample. This filter is only applicable for seqData objects.

There are no further arguments for a filter_object of type ' customFilt'.

See also

Author

Lisa Bramer, Kelly Stratton

Examples

library(pmartRdata)
to_filter <- molecule_filter(omicsData = pep_object)
summary(to_filter, min_num = 2)
#> 
#> Summary of Molecule Filter 
#> ----------------------------------
#> Minimum Number:2
#> Filtered:930
#> Not Filtered:20399
#> 
#>  num_observations frequency_counts
#>                 1              930
#>                 2             1928
#>                 3             2895
#>                 4             3826
#>                 5             4705
#>                 6             5627
#>                 7             6443
#>                 8             7213
#>                 9             7937
#>                10             8647
#>                11             9383
#>                12            10032
#>                13            10678
#>                14            11292
#>                15            11937
#>                16            12592
#>                17            13240
#>                18            13917
#>                19            14571
#>                20            15248
#>                21            15941
#>                22            16800
#>                23            17956
#>                24            21329
pep_object2 <- applyFilt(filter_object = to_filter,
                         omicsData = pep_object, min_num = 2)
summary(pep_object2) # number of Peptides is as expected based on summary of the filter object that was applied
#>                                      
#> Class                         pepData
#> Unique SampleIDs (f_data)          24
#> Unique Peptides (e_data)        20399
#> Unique RazorProteins (e_meta)    2916
#> Missing Observations           202348
#> Proportion Missing              0.413
pep_object2 <- group_designation(omicsData = pep_object2, main_effects = "Phenotype")
to_filter2 <- imdanova_filter(omicsData = pep_object2)
pep_object3 <- applyFilt(filter_object = to_filter2,
                         omicsData = pep_object2,
                         min_nonmiss_anova = 3)
#> You have specified remove_singleton_groups = TRUE, but there are no singleton groups to remove. Proceeding with application of the IMD-ANOVA filter.