Apply a S3 filter object to a pmartR S3 object

This function takes a filter object of class 'cvFilt', 'rmdFilt', 'moleculeFilt', 'proteomicsFilt', 'imdanovaFilt', or 'customFilt' and applies the filter to a dataset of pepData, proData, lipidData, or metabData.

applyFilt(filter_object, omicsData, ...)

# S3 method for moleculeFilt
applyFilt(filter_object, omicsData, min_num = 2)

# S3 method for cvFilt
applyFilt(filter_object, omicsData, cv_threshold = 150)

# S3 method for rmdFilt
applyFilt(filter_object, omicsData,
  pvalue_threshold = 1e-04)

# S3 method for proteomicsFilt
applyFilt(filter_object, omicsData,
  min_num_peps = NULL, degen_peps = FALSE)

# S3 method for imdanovaFilt
applyFilt(filter_object, omicsData,
  min_nonmiss_anova = NULL, min_nonmiss_gtest = NULL)

# S3 method for customFilt
applyFilt(filter_object, omicsData)

Arguments

filter_object
an object of the class 'cvFilt', 'proteomicsFilt', 'rmdFilt', 'moleculeFilt', 'imdanovaFilt', or 'customFilt' created by cv_filter, proteomics_filter, rmd_filter, molecule_filter, or imdanova_filter, respectively.

omicsData
an object of the class pepData, proData, lipidData, or metabData usually created by as.pepData, as.proData, as.lipidData, or as.metabData, respectively.

...
further arguments

Value

An object of the class pepData, proData, lipidData, or metabData, with specified cname_ids, edata_cnames, and emeta_cnames filtered out of the appropriate datasets.

Details

Various further arguments can be specified depending on the class of the filter_object being applied. For a filter_object of type 'moleculeFilt':

min_num

an integer value specifying the minimum number of times each biomolecule must be observed across all samples in order to retain the biomolecule. Default value is 2.

For a filter_object of type 'cvFilt':

cv_threshold

an integer value specifying the maximum coefficient of variation (CV) threshold for the biomolecules. Biomolecules with CV greater than this threshold will be filtered. Default is 150.

For a filter_object of type 'rmdFilt':

pvalue_threshold

numeric value between 0 and 1, specifying the p-value, below which samples will be removed from the dataset. Default is 0.001.

For a filter_object of type 'proteomicsFilt':

min_num_peps

an optional integer value between 1 and the maximum number of peptides that map to a protein in omicsData. The value specifies the minimum number of peptides that must map to a protein. Any protein with less than

min_num_peps	mapping to it will be returned as a protein that should be filtered. Default value is NULL.

degen_peps

logical indicator of whether to filter out degenerate peptides (TRUE) or not (FALSE). Default value is FALSE.

For a filter_object of type 'imdanovaFilt':

min_nonmiss_anova

integer value specifying the minimum number of non-missing feature values allowed per group for

anova_filter	. Default value is 2.

min_nonmiss_gtest

integer value specifying the minimum number of non-missing feature values allowed per group for

gtest_filter	. Default value is 3.

There are no futher arguments for a filter_object of type 'customFilt'.

See also

Examples

dontrun{
library(pmartRdata)
data("pep_object")
to_filter <- molecule_filter(omicsData = pep_object)
pep_object2 <- applyFilt(filter_object = to_filter, omicsData = pep_object, min_num = 2)
print(str(attributes(pep_object2)$filters))
pep_object2 <- group_designation(pep_object2, main_effects = "Condition")
to_filter2 <- imdanova_filter(omicsData = pep_object2)
pep_object3 <- applyFilt(filter_object = to_filter2, omicsData = pep_object2, min_nonmiss_anova = 3)
print(str(attributes(pep_object3)$filters))
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^