Identify Features from the Top L Order Statistics

Select features for normalization via the method of the top L order statistics (LOS)

los(e_data, edata_id, L = 0.05)

Arguments

e_data
a \(p \times n + 1\) data.frame, where \(p\) is the number of peptides, proteins, lipids, or metabolites and \(n\) is the number of samples. Each row corresponds to data for a peptide, protein, lipid, or metabolite, with a column giving the identifer name.

edata_id
character string indicating the name of the column giving the peptide, protein, lipid, or metabolite identifier. Usually obtained by calling attr(omicsData, "cnames")$edata_cname.

L
numeric value between 0 and 1, indicating the top proportion of features to be retained (default value 0.05)

Value

Character vector containing the features belonging to the subset.

Details

The feature abundances of the top L order statistics are identified and returned. Specifically, for each sample, the features with the top L proportion of highest absolute abundance are retained, and the union of these features is taken as the subset identified.

Examples

# NOT RUN {
library(pmartRdata)
data(pep_pepData)
pep_pepData2 <- group_designation(omicsData = pep_pepData, main_effects = "Condition")
pep_subset <- los(e_data = pep_pepData2$e_data, edata_id = attr(pep_pepData2, "cnames")$edata_cname, pmartR_groupDF = attr(pep_pepData2, "group_DF"))
# }