Adjust p-values for multiple comparisons

Depending upon the pval_adjust method selected, the supplied p_values are compared against an adjusted pval_thresh value or the provided means are used to compute new statistics, p-values are computed and compared against the provided pval_thresh. A data.frame that indicates which of the tests are significant, 1 if significant or 0 if insignificant. If means is also provided and the p-value is signficant then the direction of the change is indicated by the sign on 1, i.e., means<0 and p_value<pval_thresh will return -1, similarly for means>0.

p_adjustment_anova(p_values = NULL, diff_mean = NULL, t_stats = NULL,
  sizes = NULL, pval_adjust = "None")

Arguments

p_values
A matrix (or data.frame) of p-values to be adjusted.

diff_mean
A matrix (or data.frame) of groups means that are to be compared

t_stats
A matrix (or data.frame) of t-test statistics resulting from from standard procedures

sizes
A matrix (or data.frame) of group sizes

pval_adjust
character vector specifying the type of multiple comparisons adjustment to implement. A NULL value corresponds to no adjustment. Valid options include: holm, bonferroni, dunnett, tukey or none.

Value

a data frame with the following columns: group means, global G-test statistic and corresponding p-value

Examples

dontrun{
library(pmartRdata)
library(pmartR)
mypepData <- edata_transform(omicsData = pep_object, data_scale = "log2")
mypepData <- group_designation(omicsData = mypepData, main_effects = c("Condition"))
group_df <- attr(mypepData, "group_DF")
imdanova_Filt <- imdanova_filter(omicsData = mypepData)
mypepData <- applyFilt(filter_object = imdanova_Filt, omicsData = mypepData, min_nonmiss_anova=2)
anova_res <- anova_test(omicsData = mypepData)

adjusted_pvalues <- p_adjustment_anova(p_values = anova_res$Fold_change_pvalues, diff_mean = anova_res$Results[, grep("Mean", names(anova_res$Results))])
}

#> Error: <text>:1:8: unexpected '{'
#> 1: dontrun{
#>            ^